+++ CHANGES.rst
@@ -21,6 +21,7 @@ Development
 - Fix: Better explicitly initialize the ADC before reading it
 - Improve ``SystemBatteryLevel`` sensor. Thanks, `Dominik <https://github.com/ayoy>`_!
 - Make ``TelemetryTransportHTTP`` work again
+- Improve telemetry subsystem re. multi-protocol and -topology. Enable HTTP telemetry.
 
 
 2019-06-07 0.3.0